pyproject.toml
[tool.poetry]
name = "time-series-airflow-kafka-spark"
version = "0.1.0"
description = ""
authors = ["James Bristow <jbris@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<3.11"
kafka-python = "^2.0.2"
pyspark = {version = "3.5.1", extras = ["connect", "pandas-on-spark", "sql"]}
sktime = "^0.27.0"
mlflow = "^2.11.1"
numpy = "^1.26.4"
pandas = "<2.2.0"
hydra-core = "^1.3.2"
boto3 = "^1.34.59"
botocore = "^1.34.59"
matplotlib = "^3.8.3"
psycopg2 = "^2.9.9"
plotly = "^5.19.0"
prometheus-flask-exporter = "^0.23.0"
bentoml = "^1.2.6"
pydantic = "^2.6.3"
scikit-learn = "^1.4.1.post1"
scipy = "^1.12.0"
apache-airflow = {version = "2.8.2", extras = ["celery"]}
flower = "^2.0.1"
apache-airflow-providers-celery = "^3.6.1"
joblibspark = "^0.5.2"
joblib = "^1.3.2"
grpcio-status = "^1.62.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
